Move "verification modal" from Login screen to AccountVerification screen
